image/im01-introduction-image-processing/projet-ima/image-restoration/requirements.txt
numpy
jupyter
matplotlib
scikit-image
torchvision
torchaudio
imagecodecs
